Implemented realtime level/frequency readback when sweeping for SSG5000X-V (for firmware v2.x and newer)

+++ scopehal/RFSignalGenerator.cpp
@@ -447,7 +447,7 @@ void RFSignalGenerator::DoLoadConfiguration(int /*version*/, const YAML::Node& n
 		idmap.emplace(channelNode["rfgenid"].as<intptr_t>(), chan);
 
 		SetChannelOutputPower(i, channelNode["power"].as<float>());
-		SetChannelCenterFrequency(i, channelNode["centerfreq"].as<float>());
+		SetChannelCenterFrequency(i, channelNode["centerfreq"].as<double>());
 		SetChannelOutputEnable(i, channelNode["enabled"].as<bool>());
 
 		YAML::Node anode = channelNode["analogMod"];
